The numbers are like 139.00, so the Int() won't read it. Read it as Double and then convert to Int for storage

+++ DeviceCGM/CGMDictionary.swift
@@ -109,13 +109,13 @@ public class CGMDictionary: CGMAnalysis {
             return (-1, .none)
         }
 
-        guard let value = Int(elements[5]) else {
+        guard let value = Double(elements[5]) else {
             return (-1, .none)
         }
         
         return (ptId, Record(
             timestamp: elements[3],
-            value: value
+            value: Int(value)
         ))
     }
 }
